refactor(imports): target facade implementations directly
