+++ scripts/provider_balances.py
@@ -20,7 +20,8 @@ sys.path.insert(0, str(Path(__file__).resolve().parent.parent / "src"))
 
 from sqlalchemy.exc import OperationalError  # noqa: E402
 
-from treg import ledger, reconcile  # noqa: E402
+from treg import reconcile  # noqa: E402
+from treg.domain import money as ledger  # noqa: E402
 from treg.config import get_settings  # noqa: E402
 from treg.infra.db import session_maker  # noqa: E402
 from treg.domain.capacity.collectors import (  # noqa: E402,F401 — re-exported for older callers
