+++ src/treg/api.py
@@ -26,9 +26,10 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import defer
 from sqlmodel import select
 
-from . import audit, catalog_store, crypto, localrun, ratestore, runner, sandbox as demo_sandbox
+from . import audit, crypto, localrun, ratestore, runner, sandbox as demo_sandbox
 from .caller_metadata import _client_of
 from .config import get_settings
+from .domain.catalog import store as catalog_store
 from .infra.db import get_session, session_maker
 from .domain.governance.teams import _unique_slug
 from .domain.identity.access import (
