+++ src/treg/application/signup.py
@@ -8,10 +8,12 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlmodel import select
 
-from .. import adsconv, health, ledger, referrals, sandbox as demo_sandbox
-from ..infra.db import session_maker
+from .. import adsconv, health, sandbox as demo_sandbox
+from ..domain import money as ledger
+from ..domain import referrals
 from ..domain.governance.teams import _make_org_membership, _slugify
 from ..domain.identity.access import _is_machine_email, _norm_email
+from ..infra.db import session_maker
 from ..models import Org, User
 from ..timeutil import utcnow_naive as _utcnow_naive
