+++ src/treg/archive.py
@@ -270,7 +270,7 @@ async def _store(
         from sqlalchemy import select
         from sqlalchemy.exc import IntegrityError
 
-        from . import catalog_store
+        from .domain.catalog import store as catalog_store
         from .infra.db import session_maker
         from .models import ArchiveKey, ArchiveSnapshot
 
@@ -429,7 +429,7 @@ async def lookup(
             return None
         from sqlalchemy import select
 
-        from . import catalog_store
+        from .domain.catalog import store as catalog_store
         from .infra.db import session_maker
         from .models import ArchiveKey, ArchiveSnapshot
 
@@ -639,7 +639,7 @@ async def refresh_once(client) -> int:
         return 0
     from sqlalchemy import func, select
 
-    from . import catalog_store
+    from .domain.catalog import store as catalog_store
     from .infra.db import session_maker
     from .models import ArchiveKey, ArchiveSnapshot
