+++ src/treg/mcp.py
@@ -51,7 +51,8 @@ from mcp.server.transport_security import TransportSecuritySettings
 from mcp.shared.exceptions import MCPError
 from mcp.types import METHOD_NOT_FOUND, ToolAnnotations
 
-from . import audit, catalog_store
+from . import audit
+from .domain.catalog import store as catalog_store
 from .config import PUBLIC_HOST_ALIASES, get_settings
 from .domain.catalog.stats import EndpointObservationReader
 
@@ -1368,7 +1369,7 @@ class RequireAuthForProtectedTools:
         return None             # a live access token, or a per-org token the tool validates itself
 
     async def _challenge(self, send, *, invalid: bool = False) -> None:
-        from . import mcp_oauth
+        from .domain.identity import mcp_oauth
 
         base = get_settings().public_url.rstrip("/")
         suffix = "/mcp/v2" if self.resource_version == "v2" else ""
