+++ src/treg/routers/billing.py
@@ -4,10 +4,11 @@ from fastapi import APIRouter, Depends, HTTPException, Query, Request
 from pydantic import BaseModel
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from .. import billing, ledger
+from ..application import billing
 from ..config import get_settings
-from ..infra.db import get_session
+from ..domain import money as ledger
 from ..domain.identity.access import Caller, _role_at_least, require_member
+from ..infra.db import get_session
 from ..models import Org
 from .auth_helpers import _is_https
 from .orgs import _require_admin_of
