+++ src/treg/routers/call.py
@@ -9,7 +9,7 @@ from fastapi.responses import Response, StreamingResponse
 from starlette.background import BackgroundTask
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from .. import audit, catalog_store, ledger, oauth_providers
+from .. import audit, oauth_providers
 from .. import sandbox as demo_sandbox
 from ..application.call.idempotency import (
     _release_idempotent_claim as release_idempotent_claim,
@@ -33,10 +33,12 @@ from ..application.call.intake import (
 from ..application.call.types import CallerSnapshot, CallFailure, CallInput, UpstreamResponse
 from ..caller_metadata import _client_of
 from ..config import get_settings
-from ..infra.db import get_session
+from ..domain import money as ledger
+from ..domain.catalog import store as catalog_store
 from ..domain.governance import access as access_policy
 from ..domain.governance import publicdemo as publicdemo_policy
 from ..domain.identity.access import Caller, require_member
+from ..infra.db import get_session
 from ..models import Tool
 from .auth import _client_ip
 from .orgs import count_today
