+++ src/treg/routers/web.py
@@ -17,7 +17,9 @@ from fastapi.responses import (FileResponse, HTMLResponse, JSONResponse, PlainTe
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlmodel import select
 
-from .. import adsconv, agent_pages, catalog_store, oauth_providers, referrals
+from .. import adsconv, agent_pages, oauth_providers
+from ..domain import referrals
+from ..domain.catalog import store as catalog_store
 from ..domain.identity import session as sess
 from ..config import PUBLIC_HOST_ALIASES, get_settings
 from ..infra.db import get_session
