+++ tests/callmatrix/test_matrix.py
@@ -12,9 +12,10 @@ from sqlalchemy.exc import TimeoutError as PoolTimeoutError
 
 from treg.application.call import service as call_service
 from treg.routers import call as call_routes
-from treg import audit, ledger
+from treg import audit
+from treg.domain import money as ledger
 from treg.infra.db import session_maker
-from treg.ledger import with_margin
+from treg.domain.money import with_margin
 from treg.models import CallRecord
 
 from test_marketplace_call import (
