+++ tests/test_auth.py
@@ -11,7 +11,8 @@ from fastapi import FastAPI
 from httpx import ASGITransport, AsyncClient
 from sqlmodel import select
 
-from treg import crypto, session as sess
+from treg import crypto
+from treg.domain.identity import session as sess
 from treg.api import app
 from treg.config import get_settings
 from treg.infra.db import reset_db, session_maker
