+++ tests/test_bughunt_server.py
@@ -15,7 +15,8 @@ from fastapi import FastAPI, Request
 from httpx import ASGITransport, AsyncClient
 from sqlmodel import select
 
-from treg import audit, crypto, oauth, session as sess
+from treg import audit, crypto, oauth
+from treg.domain.identity import session as sess
 from treg.api import app
 from treg.config import get_settings
 from treg.infra.db import reset_db, session_maker
