+++ tests/test_call_pool_discipline.py
@@ -29,7 +29,8 @@ from treg import api as A
 from treg.domain.catalog import stats as catalog_stats
 from treg.application.call import service as call_service
 from treg.routers import call as call_routes
-from treg import audit, ledger
+from treg import audit
+from treg.domain import money as ledger
 from treg.config import get_settings
 from treg.infra.db import _engine, session_maker
 from treg.infra.upstream.relay import relay as upstream_relay
