+++ tests/test_ledger_claim_review.py
@@ -7,7 +7,7 @@ from httpx import AsyncClient
 from sqlalchemy import delete
 from sqlmodel import select
 
-from treg import ledger
+from treg.domain import money as ledger
 from treg.infra.db import session_maker
 from treg.models import Hold, LedgerEntry
