+++ tests/test_login_page.py
@@ -8,7 +8,8 @@ from __future__ import annotations
 import pytest
 from httpx import ASGITransport, AsyncClient
 
-from treg import crypto, session as sess
+from treg import crypto
+from treg.domain.identity import session as sess
 from treg.api import app
 from treg.config import get_settings
 from treg.infra.db import reset_db, session_maker
