+++ tests/test_single_user.py
@@ -10,7 +10,8 @@ import pytest
 from httpx import ASGITransport, AsyncClient
 from sqlmodel import select
 
-from treg import api, maintenance, session as sess
+from treg import api, maintenance
+from treg.domain.identity import session as sess
 from treg.__main__ import _prepare_serve
 from treg.api import LOCAL_ORG_NAME, LOCAL_USER_EMAIL, app
 from treg.config import Settings, get_settings
