+++ tests/test_tag_billing.py
@@ -17,7 +17,8 @@ import pytest
 from httpx import AsyncClient
 from sqlmodel import select
 
-from treg import audit, crypto, ledger
+from treg import audit, crypto
+from treg.domain import money as ledger
 from treg.application.call import service as call_service
 from treg.application.call import settle as call_settle
 from treg.application.call.types import UpstreamResponse
